[스프링 - 협업 미션] 삭정(성정석) 미션 제출합니다. (#5)
* feat: 요금 조회 기능 구현

* refactor: 설정 파일 분리

* fix: mysql conneciton

* fix: db initialization 가능하게 고침

* fix: jaxb-osgi 에러 관련 문제

* fix: bean dependecny 에러 해결1

* fix: bean dependecny 에러 해결2

* feat: dataloader 테스트 에만 적용 되도록 수정

* fix: lineResponse 바인딩 문제 해결

* feat: 요금 정책 추가 초안

* feat: 역 이름, 구간거리 수정 기능 추가

* feat: 구간 정보 조회 기능 구현

* fix: login 이슈 해결

* feat: stations 조회시 line 정보도 추가해서 보내도록 수정

* feat: controller advice 추가

* feat: error handling 추가

* fix: 서비스 순환참조 문제 해결

* fix: application.yml 설정파일 변경

* refactor: error handling 표준 통일을 위해 nosuchEmailException 제거

* fix: erroDto getter 추가

* feat: 지하철 지도를 위한 전체라인 조회 기능 추가

* feat: Fare 객체 추가

* feat: Fare와 관련된 enum 추가

* refactor: distance 객체 생성

* refactor: 도메인 객체 수정

* test: 요금 조회 테스트 추가

* refactor: Age 객체 추가

* refactor: 코드 리포맷팅 및 변수명 수정

* feat: 값 객체 validation 추가

* fix: distance validation 제거

지하철 지도를 위한 조회시 마지막 역의 distance는 -1이 되어야 하므로

* fix: controller 내부의 불필요한 exception handler 제거

* fix: lineDao query case 문제 해결

* feat: MapController CORS 설정 추가

* feat: stack trace출력하도록 하는 기능 추가

* fix: 구간 조회 에러 해결

db query로 처리하던 부분 도메인 로직으로 대체

* fix: line 에서 extraFare를 디폴트값으로 들고 있도록 리팩토링

* fix: line 삭제시 관련 section 들도 자동으로 삭제되도록 수정

* feat: 구간 정렬 기능 추가

* feat: 회원가입시 validation기능 추가

* fix: minimum age 조건 수정

* refactor: exception들 리팩토링

* feat: 로그인 인증과 관련된 예외 처리 세분화

* docs(gitignore): build_script.sh 파일 무시하지 않도록 설정

* feat: stationRequest validation 기능 추가

* feat: lineService 관련 더 세부적인 예외 처리 추가

* refactor: 추가 예외 처리에 따른 테스트 코드 수정

* feat: swagger 기능 추가

* feat: swagger models 부분에 request dto의 validation 사항도 나오게끔 하는 기능 추가

* refactor: 전체 역을 조회하는 로직 리팩토링

* refactor: lineDao stationDao에서 의미있는 exception을 던지도록 리팩토링

* refactor: discount enum 내부 리팩토링

* refactor: distance와 fare 구하는 로직 리팩토링

디미터의 법칙 준수

* refactor: 예외를 던지지 않는 새로운 argument resolver 추가

* feat: section 생성시 상행역 하행역 동일성 검증 도메인 객체에 추가

* fix: distanceCharge에서 enum들 순서가 바뀌더라도 로직에 이상이 없도록 수정

* test: Distance관련 테스트 추가

* test: 요금 정책 관련 도메인들에 대한 테스트 추가

* feat: logging 기능 추가

* fix: gradle lombok annotation 에러 관련 dependency 추가

* docs(gitignore): renew git ignore

* refactor: 인수 테스트에서 공통적으로 사용되는 로직들 acceptanceTest로 이동

* refactor: test 코드 리팩토링

코드 중복 정리 및 메서드 네이밍 변경

* refactor: domain test 추가 리팩토링

* style: 코드 리포맷팅

+++ src/main/java/wooteco/subway/line/application/LineService.java
@@ -1,24 +1,28 @@
 package wooteco.subway.line.application;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import wooteco.subway.line.dao.LineDao;
 import wooteco.subway.line.dao.SectionDao;
 import wooteco.subway.line.domain.Line;
 import wooteco.subway.line.domain.Section;
-import wooteco.subway.line.dto.LineRequest;
-import wooteco.subway.line.dto.LineResponse;
-import wooteco.subway.line.dto.SectionRequest;
+import wooteco.subway.line.domain.Sections;
+import wooteco.subway.line.dto.*;
+import wooteco.subway.path.domain.Distance;
+import wooteco.subway.path.domain.Fare;
 import wooteco.subway.station.application.StationService;
 import wooteco.subway.station.domain.Station;
+import wooteco.subway.station.dto.StationMapResponse;
+import wooteco.subway.station.dto.StationTransferResponse;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
 @Service
 public class LineService {
-    private LineDao lineDao;
-    private SectionDao sectionDao;
-    private StationService stationService;
+    private final LineDao lineDao;
+    private final SectionDao sectionDao;
+    private final StationService stationService;
 
     public LineService(LineDao lineDao, SectionDao sectionDao, StationService stationService) {
         this.lineDao = lineDao;
@@ -26,8 +30,13 @@ public class LineService {
         this.stationService = stationService;
     }
 
+    @Transactional
     public LineResponse saveLine(LineRequest request) {
-        Line persistLine = lineDao.insert(new Line(request.getName(), request.getColor()));
+        Line persistLine = lineDao.insert(new Line(request.getName(), request.getColor(), new Fare(request.getExtraFare())));
+
+        if (request.getUpStationId().equals(request.getDownStationId())) {
+            throw new SameUpAndDownStationException();
+        }
         persistLine.addSection(addInitSection(persistLine, request));
         return LineResponse.of(persistLine);
     }
@@ -36,7 +45,7 @@ public class LineService {
         if (request.getUpStationId() != null && request.getDownStationId() != null) {
             Station upStation = stationService.findStationById(request.getUpStationId());
             Station downStation = stationService.findStationById(request.getDownStationId());
-            Section section = new Section(upStation, downStation, request.getDistance());
+            Section section = new Section(upStation, downStation, new Distance(request.getDistance()));
             return sectionDao.insert(line, section);
         }
         return null;
@@ -59,28 +68,40 @@ public class LineService {
     }
 
     public Line findLineById(Long id) {
+        validateLineExistence(id);
         return lineDao.findById(id);
     }
 
+    @Transactional
     public void updateLine(Long id, LineRequest lineUpdateRequest) {
+        validateLineExistence(id);
         lineDao.update(new Line(id, lineUpdateRequest.getName(), lineUpdateRequest.getColor()));
     }
 
+    @Transactional
     public void deleteLineById(Long id) {
+        validateLineExistence(id);
         lineDao.deleteById(id);
     }
 
+    @Transactional
     public void addLineStation(Long lineId, SectionRequest request) {
+        validateLineExistence(lineId);
+        if (request.getUpStationId().equals(request.getDownStationId())) {
+            throw new SameUpAndDownStationException();
+        }
         Line line = findLineById(lineId);
         Station upStation = stationService.findStationById(request.getUpStationId());
         Station downStation = stationService.findStationById(request.getDownStationId());
-        line.addSection(upStation, downStation, request.getDistance());
+        line.addSection(upStation, downStation, new Distance(request.getDistance()));
 
         sectionDao.deleteByLineId(lineId);
         sectionDao.insertSections(line);
     }
 
+    @Transactional
     public void removeLineStation(Long lineId, Long stationId) {
+        validateLineExistence(lineId);
         Line line = findLineById(lineId);
         Station station = stationService.findStationById(stationId);
         line.removeSection(station);
@@ -89,4 +110,87 @@ public class LineService {
         sectionDao.insertSections(line);
     }
 
+    @Transactional
+    public void updateDistance(final long lineId, final long upStationId, final long downStationId, final Integer distance) {
+        if (sectionDao.doesSectionNotExists(lineId, upStationId, downStationId)) {
+            throw new NoSuchSectionException();
+        }
+        sectionDao.updateDistance(lineId, upStationId, downStationId, distance);
+    }
+
+    public LineSectionResponse findSectionsByLineId(final long lineId) {
+        List<Line> lines = lineDao.findAll();
+
+        Line line = getLine(lines, lineId);
+        List<StationTransferResponse> stationsWithTransferLines = getStationsWithTransferLines(line, lines);
+        List<SectionResponse> sectionResponses = SectionResponse.listOf(line.getSections());
+
+        return LineSectionResponse.of(
+                line.getId(),
+                line.getName(),
+                line.getColor(),
+                stationsWithTransferLines,
+                sectionResponses
+        );
+    }
+
+    private Line getLine(final List<Line> lines, final long lineId) {
+        return lines.stream()
+                .filter(line -> line.hasId(lineId))
+                .findFirst()
+                .orElseThrow(NoSuchLineException::new);
+    }
+
+    private List<StationTransferResponse> getStationsWithTransferLines(final Line line, final List<Line> lines) {
+        return line.getStations().stream()
+                .map(station -> {
+                    List<TransferLineResponse> transferLineResponses = getTransferLineResponses(station, line.getId(), lines);
+                    return new StationTransferResponse(station.getId(), station.getName(), transferLineResponses);
+                })
+                .collect(Collectors.toList());
+    }
+
+    private List<TransferLineResponse> getTransferLineResponses(final Station station, final long lineId, final List<Line> lines) {
+        return lines.stream()
+                .filter(line -> line.contains(station) && !line.hasId(lineId))
+                .map(line -> new TransferLineResponse(line.getId(), line.getName(), line.getColor()))
+                .collect(Collectors.toList());
+    }
+
+    public void validateLineExistence(final Long id) {
+        if (lineDao.doesIdNotExist(id)) {
+            throw new NoSuchLineException();
+        }
+    }
+
+    public List<LineMapResponse> findLinesForMap() {
+        List<Line> lines = lineDao.findAll();
+
+        return lines.stream()
+                .map(line -> {
+                    List<StationMapResponse> stationMapResponses = getStationMapResponses(line, lines);
+                    return new LineMapResponse(line.getId(), line.getName(), line.getColor(), stationMapResponses);
+                })
+                .collect(Collectors.toList());
+    }
+
+    private List<StationMapResponse> getStationMapResponses(final Line line, final List<Line> lines) {
+        Sections sections = line.getSections();
+        List<Station> stations = sections.getStations();
+
+        return stations.stream()
+                .map(station -> {
+                    Distance distanceToNextStation = sections.getDistanceToNextStation(station);
+                    List<TransferLineResponse> transferLineResponses = getTransferLineResponses(lines, line, station);
+                    return new StationMapResponse(station.getId(), station.getName(), distanceToNextStation.getDistance(), transferLineResponses);
+                })
+                .collect(Collectors.toList());
+    }
+
+    private List<TransferLineResponse> getTransferLineResponses(final List<Line> lines, final Line currentLine, final Station station) {
+        return lines.stream()
+                .filter(line -> line.contains(station) && !line.equals(currentLine))
+                .map(transferLine -> new TransferLineResponse(transferLine.getId(), transferLine.getName(), transferLine.getColor()))
+                .collect(Collectors.toList());
+    }
 }
